yoyoyoyoyo

+++ flask/Unit-01/07-sql-alchemy-2/app.py
@@ -2,12 +2,15 @@ from flask import Flask, request, redirect, url_for, render_template
 from flask_modus import Modus
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from forms import UserForm, MessageForm, DeleteForm
+import os
 
 app = Flask(__name__)
 app.config[
     'SQLALCHEMY_DATABASE_URI'] = "postgres://localhost/flask_sql_alchemy2"
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SQLALCHEMY_ECHO'] = True
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
 modus = Modus(app)
 db = SQLAlchemy(app)
 Migrate(app, db)
@@ -39,64 +42,100 @@ def root():
 
 @app.route('/users', methods=['GET', 'POST'])
 def index():
+    delete_form = DeleteForm()
     if request.method == "POST":
-        new_user = User(
-            first_name=request.form['first_name'],
-            last_name=request.form['last_name'])
-        db.session.add(new_user)
-        db.session.commit()
-        return redirect(url_for('index'))
-    return render_template('/users/index.html', users=User.query.all())
+        form = UserForm(request.form)
+        if form.validate():
+            new_user = User(
+                first_name=request.form['first_name'],
+                last_name=request.form['last_name'])
+            db.session.add(new_user)
+            db.session.commit()
+            return redirect(url_for('index'))
+        else:
+            return render_template('users/new.html', form=form)
+    return render_template(
+        '/users/index.html', users=User.query.all(), delete_form=delete_form)
 
 
 @app.route('/users/new')
 def new():
-    return render_template('users/new.html')
+    user_form = UserForm()
+    return render_template('users/new.html', form=user_form)
 
 
 @app.route('/users/<int:id>/edit')
 def edit(id):
-    return render_template('/users/edit.html', user=User.query.get(id))
+    delete_form = DeleteForm()
+    user_form = UserForm(obj=User.query.get(id))
+    return render_template(
+        '/users/edit.html',
+        user=User.query.get(id),
+        form=user_form,
+        delete_form=delete_form)
 
 
 @app.route('/users/<int:id>', methods=['GET', 'PATCH', 'DELETE'])
 def show(id):
+    delete_form = DeleteForm(request.form)
     found_user = User.query.get(id)
+    form = UserForm(request.form)
     if request.method == b'PATCH':
-        found_user.first_name = request.form['first_name']
-        found_user.last_name = request.form['last_name']
-        db.session.add(found_user)
-        db.session.commit()
-        return redirect(url_for('index'))
+        if form.validate():
+            found_user.first_name = request.form['first_name']
+            found_user.last_name = request.form['last_name']
+            db.session.add(found_user)
+            db.session.commit()
+            return redirect(url_for('index'))
+        else:
+            render_template(
+                '/users/edit.html',
+                user=found_user,
+                form=form,
+                delete_form=delete_form)
     if request.method == b'DELETE':
-        db.session.delete(found_user)
-        db.session.commit()
+        if delete_form.validate():
+            db.session.delete(found_user)
+            db.session.commit()
         return redirect(url_for('index'))
     return render_template('/users/show.html', user=found_user)
 
 
 @app.route('/users/<int:user_id>/messages/index', methods=['GET', 'POST'])
 def messages_index(user_id):
+    delete_form = DeleteForm(request.form)
     if request.method == 'POST':
-        new_message = Message(content=request.form['content'], user_id=user_id)
-        db.session.add(new_message)
-        db.session.commit()
-        return redirect(url_for('messages_index', user_id=user_id))
+        form = MessageForm(request.form)
+        if form.validate():
+            new_message = Message(
+                content=request.form['content'], user_id=user_id)
+            db.session.add(new_message)
+            db.session.commit()
+            return redirect(url_for('messages_index', user_id=user_id))
+        return render_template(
+            '/messages/edit.html', user_id=user_id, form=form)
     return render_template(
-        '/messages/index.html', user=User.query.get(user_id))
+        '/messages/index.html',
+        user=User.query.get(user_id),
+        delete_form=delete_form)
 
 
 @app.route('/users/<int:user_id>/messages/new')
 def messages_new(user_id):
-    return render_template('/messages/new.html', user_id=user_id)
+    message_form = MessageForm()
+    return render_template(
+        '/messages/new.html', user_id=user_id, form=message_form)
 
 
 @app.route('/users/<int:user_id>/messages/<int:id>/edit')
 def messages_edit(user_id, id):
+    found_message = Message.query.get(id)
+    form = MessageForm(obj=found_message)
     return render_template(
         '/messages/edit.html',
         user=User.query.get(user_id),
-        message=Message.query.get(id))
+        message=found_message,
+        form=form)
 
 
 @app.route(
@@ -104,14 +143,33 @@ def messages_edit(user_id, id):
     methods=["GET", "PATCH", "DELETE"])
 def messages_show(user_id, id):
     found_message = Message.query.get(id)
+    delete_form = DeleteForm(request.form)
     if request.method == b"PATCH":
-        found_message.content = request.form["content"]
-        db.session.add(found_message)
-        db.session.commit()
-        return redirect(url_for('messages_index', user_id=user_id))
+        form = MessageForm(request.form)
+        if form.validate():
+            found_message.content = request.form["content"]
+            db.session.add(found_message)
+            db.session.commit()
+            return redirect(url_for('messages_index', user_id=user_id, id=id))
+        else:
+            return render_template(
+                '/messages/edit.html',
+                user_id=user_id,
+                message=found_message,
+                form=form)
+        return redirect(url_for('messages_index', user_id=user_id, id=id))
     if request.method == b'DELETE':
-        db.session.delete(found_message)
-        db.session.commit()
-        return redirect(url_for('messages_index', user_id=user_id))
+        if delete_form.validate():
+            db.session.delete(found_message)
+            db.session.commit()
+        return redirect(
+            url_for(
+                'messages_index',
+                user_id=user_id,
+                delete_form=delete_form,
+                id=id))
     return render_template(
-        '/messages/show.html', message=found_message, user=user_id)
+        '/messages/show.html',
+        message=found_message,
+        user=user_id,
+        delete_form=delete_form)
